+++ src/test/java/com/atwoz/mission/application/membermission/MemberMissionsServiceTest.java
@@ -5,8 +5,8 @@ import com.atwoz.mission.domain.membermission.MemberMissions;
 import com.atwoz.mission.domain.membermission.MemberMissionsRepository;
 import com.atwoz.mission.domain.mission.Mission;
 import com.atwoz.mission.domain.mission.MissionRepository;
-import com.atwoz.mission.exception.MissionNotClearException;
-import com.atwoz.mission.exception.MissionNotFoundException;
+import com.atwoz.mission.exception.mission.exceptions.MissionNotClearException;
+import com.atwoz.mission.exception.mission.exceptions.MissionNotFoundException;
 import com.atwoz.mission.exception.membermission.exceptions.MemberMissionNotFoundException;
 import com.atwoz.mission.exception.membermission.exceptions.MemberMissionsNotFoundException;
 import com.atwoz.mission.intrastructure.membermission.MemberMissionsFakeRepository;
