+++ src/test/java/com/atwoz/mission/application/mission/MissionServiceTest.java
@@ -3,11 +3,10 @@ package com.atwoz.mission.application.mission;
 import com.atwoz.mission.application.mission.dto.MissionCreateRequest;
 import com.atwoz.mission.domain.mission.Mission;
 import com.atwoz.mission.domain.mission.MissionRepository;
-import com.atwoz.mission.exception.MissionTypeInvalidException;
-import com.atwoz.mission.exception.PublicOptionInvalidException;
-import com.atwoz.mission.exception.RewardValueInvalidException;
+import com.atwoz.mission.exception.mission.exceptions.MissionTypeInvalidException;
+import com.atwoz.mission.exception.mission.exceptions.PublicOptionInvalidException;
+import com.atwoz.mission.exception.mission.exceptions.RewardValueInvalidException;
 import com.atwoz.mission.intrastructure.mission.MissionFakeRepository;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
